Hooking up the adapter to Vault queries. Adjusting price calculation. Test adjustment and cleanup
