+++ test/lp/constants.ts
@@ -33,6 +33,16 @@ export const BOOSTED_POOL_ADDRESS =
 export const BOOSTED_GAUGE_ADDRESS =
   "0x68d019f64A7aa97e2D4e7363AEE42251D08124Fb";
 
+export const BOOSTED_GAUGE_TOP_HOLDERS = [
+  "0x995a09ed0b24ee13fbfcfbe60cad2fb6281b479f",
+  "0xb1ff8bf9c3a55877b5ee38e769e7a78cd000848e",
+  "0x44e5f536429363dd2a20ce31e3666c300233d151",
+  "0x81fa0f35b54790f78e76c74d05bd6d95632c030b",
+  "0x3a3ee61f7c6e1994a2001762250a5e17b2061b6d",
+  "0x8d8f55c99971b9e69967a406419ed815ee63d3cd",
+  "0x98bea99727b297f5eca448d1640075f349c08547",
+];
+
 export const VAULT_ADDRESS = "0xBA12222222228d8Ba445958a75a0704d566BF2C8";
 export const HELPERS_ADDRESS = "0x5aDDCCa35b7A0D07C74063c48700C8590E87864E";
